Update Frontend Formatter (#1335)
This PR changes some configs of the frontend formatter (`Prettier`):

- `printWidth`: 80 -> 120,
- `trailingComma`: "none" ->  "es5",
- `arrowParens`: "aways" -> "avoid",

+++ core/new-gui/decorate-angular-cli.js
@@ -48,9 +48,8 @@ function symlinkNgCLItoNxCLI() {
        * This is the most reliable way to create symlink-like behavior on Windows.
        * Such that it works in all shells and works with npx.
        */
-      ["", ".cmd", ".ps1"].forEach((ext) => {
-        if (fs.existsSync(nxPath + ext))
-          fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
+      ["", ".cmd", ".ps1"].forEach(ext => {
+        if (fs.existsSync(nxPath + ext)) fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
       });
     } else {
       // If unix-based, symlink
@@ -58,9 +57,7 @@ function symlinkNgCLItoNxCLI() {
     }
   } catch (e) {
     output.error({
-      title:
-        "Unable to create a symlink from the Angular CLI to the Nx CLI:" +
-        e.message
+      title: "Unable to create a symlink from the Angular CLI to the Nx CLI:" + e.message,
     });
     throw e;
   }
@@ -70,10 +67,10 @@ try {
   symlinkNgCLItoNxCLI();
   require("@nrwl/cli/lib/decorate-cli").decorateCli();
   output.log({
-    title: "Angular CLI has been decorated to enable computation caching."
+    title: "Angular CLI has been decorated to enable computation caching.",
   });
 } catch (e) {
   output.error({
-    title: "Decoration of the Angular CLI did not complete successfully"
+    title: "Decoration of the Angular CLI did not complete successfully",
   });
 }
